pyproject.toml
[tool.poetry]
name = "store-api"
version = "0.1.0"
description = ""
authors = ["Anton Pashinov <troxin1988@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
pyjwt = "^2.9.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = "src/migrations"